+++ elastic-job-lite/elastic-job-lite-core/src/main/java/com/dangdang/ddframe/job/lite/internal/failover/FailoverListenerManager.java
@@ -21,10 +21,11 @@ import com.dangdang.ddframe.job.lite.config.LiteJobConfiguration;
 import com.dangdang.ddframe.job.lite.internal.config.ConfigurationNode;
 import com.dangdang.ddframe.job.lite.internal.config.ConfigurationService;
 import com.dangdang.ddframe.job.lite.internal.config.LiteJobConfigurationGsonFactory;
+import com.dangdang.ddframe.job.lite.internal.instance.InstanceNode;
 import com.dangdang.ddframe.job.lite.internal.listener.AbstractJobListener;
 import com.dangdang.ddframe.job.lite.internal.listener.AbstractListenerManager;
+import com.dangdang.ddframe.job.lite.internal.schedule.JobRegistry;
 import com.dangdang.ddframe.job.lite.internal.sharding.ExecutionService;
-import com.dangdang.ddframe.job.lite.internal.sharding.ShardingNode;
 import com.dangdang.ddframe.job.lite.internal.sharding.ShardingService;
 import com.dangdang.ddframe.job.reg.base.CoordinatorRegistryCenter;
 import org.apache.curator.framework.recipes.cache.TreeCacheEvent.Type;
@@ -36,6 +37,8 @@ import org.apache.curator.framework.recipes.cache.TreeCacheEvent.Type;
  */
 public final class FailoverListenerManager extends AbstractListenerManager {
     
+    private final String jobName;
+    
     private final ConfigurationService configService;
     
     private final ExecutionService executionService;
@@ -46,18 +49,19 @@ public final class FailoverListenerManager extends AbstractListenerManager {
     
     private final ConfigurationNode configNode;
     
-    private final ShardingNode shardingNode;
+    private final InstanceNode instanceNode;
     
     private final FailoverNode failoverNode;
     
     public FailoverListenerManager(final CoordinatorRegistryCenter regCenter, final String jobName) {
         super(regCenter, jobName);
+        this.jobName = jobName;
         configService = new ConfigurationService(regCenter, jobName);
         executionService = new ExecutionService(regCenter, jobName);
         shardingService = new ShardingService(regCenter, jobName);
         failoverService = new FailoverService(regCenter, jobName);
         configNode = new ConfigurationNode(jobName);
-        shardingNode = new ShardingNode(jobName);
+        instanceNode = new InstanceNode(jobName);
         failoverNode = new FailoverNode(jobName);
     }
     
@@ -68,27 +72,25 @@ public final class FailoverListenerManager extends AbstractListenerManager {
         addDataListener(new FailoverSettingsChangedJobListener());
     }
     
-    private void failover(final Integer item, final Type eventType) {
-        if (!isJobCrashAndNeedFailover(item, eventType)) {
-            return;
-        }
-        failoverService.setCrashedFailoverFlag(item);
-        if (!executionService.hasRunningItems(shardingService.getLocalShardingItems())) {
-            failoverService.failoverIfNecessary();
-        }
-    }
-    
-    private boolean isJobCrashAndNeedFailover(final Integer item, final Type eventType) {
+    private boolean isFailoverEnabled() {
         LiteJobConfiguration jobConfig = configService.load(true);
-        boolean isFailover = null != jobConfig && jobConfig.isFailover();
-        return null != item && Type.NODE_REMOVED == eventType && !executionService.isCompleted(item) && isFailover;
+        return null != jobConfig && jobConfig.isFailover();
     }
     
     class JobCrashedJobListener extends AbstractJobListener {
         
         @Override
         protected void dataChanged(final String path, final Type eventType, final String data) {
-            failover(shardingNode.getItemByRunningItemPath(path), eventType);
+            if (isFailoverEnabled() && Type.NODE_REMOVED == eventType && instanceNode.isInstancePath(path)) {
+                String jobInstanceId = path.substring(instanceNode.getInstanceFullPath().length() + 1);
+                if (jobInstanceId.equals(JobRegistry.getInstance().getJobInstance(jobName).getJobInstanceId())) {
+                    return;
+                }
+                for (int each : shardingService.getShardingItems(jobInstanceId)) {
+                    failoverService.setCrashedFailoverFlag(each);
+                    failoverService.failoverIfNecessary();
+                }
+            }
         }
     }
     
@@ -96,7 +98,17 @@ public final class FailoverListenerManager extends AbstractListenerManager {
         
         @Override
         protected void dataChanged(final String path, final Type eventType, final String data) {
-            failover(failoverNode.getItemByExecutionFailoverPath(path), eventType);
+            if (isFailoverEnabled() && Type.NODE_REMOVED == eventType) {
+                Integer item = failoverNode.getItemByExecutionFailoverPath(path);
+                if (isJobCrashed(item)) {
+                    failoverService.setCrashedFailoverFlag(item);
+                    failoverService.failoverIfNecessary();
+                }
+            }
+        }
+        
+        private boolean isJobCrashed(final Integer item) {
+            return null != item && !executionService.isCompleted(item);
         }
     }