+++ elastic-job-doc/elastic-job-lite-doc/content/00-overview/release-notes.md
@@ -7,6 +7,13 @@ prev = "/00-overview/intro/"
 next = "/00-overview/contribution/"
 +++
 
+## 2.1.2
+
+### 功能提升
+
+1. [ISSUE #296](https://github.com/dangdangdotcom/elastic-job/issues/296) Elastic-Job-Cloud运维界面重构
+1. [ISSUE #302](https://github.com/dangdangdotcom/elastic-job/issues/302) Failover与MonitorExecution分离
+
 ## 2.1.1
 
 ### 新功能